… missing from previous commit (interface update for new param input) 	- note, also need to update ABI in smartcontractgui.xyz

+++ contracts/ICallitLib.sol
@@ -91,7 +91,7 @@ interface ICallitLib {
     function genHashOfAddies(address[] calldata addies) external pure returns (address);
     function _generateAddressHash(address host, string memory uid) external pure returns (address);
     function _getWinningVoteIdxForMarket(uint64[] memory _resultTokenVotes) external view returns(uint16);
-    function _perc_of_uint64(uint32 _perc, uint64 _num) external pure returns (uint64);
+    function _perc_of_uint64(uint16 _perc, uint64 _num) external pure returns (uint64);
     function _deductFeePerc(uint64 _net_usdAmnt, uint16 _feePerc, uint64 _usdAmnt) external pure returns(uint64);
     function _uint64_from_uint256(uint256 value) external pure returns (uint64);
     function _perc_total_supply_owned(address _token, address _account) external view returns (uint64);
